Evaluate all entries for polynomial basis functions in `h_eval`
The current behavior of `h_eval` is to only evaluate entries
where the newx point is greater than the knot `t_k` for the kth
basis function.  This makes sense for the FF basis functions,
but the polynomial basis functions are nonzero for `x < t_k`
(except at the design points, which are roots).