Include the dgCMatrix class in dspline namespace
The dgCMatrix class must be available in the program environment for the
matrix construction and evaluation functions to work.

In the previous implementation, the call of `Matrix::sparseMatrix`
implicitly made the class available because the `::` operator
implicitly loads a namespace (even though it does not attach it).
Because no such call to a `Matrix` package function is present now,
we instead explicitly include dgCMatrix in the dspline namespace,
so that when dspline is attached, so is the dgCMatrix class.

Another solution is to promote the Matrix package from an Imports
to a Depends; this would attach everything from the Matrix package
whenever the dspline package is attached.  But, that is less clean
and generally discouraged.

See also:
* https://cran.r-project.org/web/packages/roxygen2/vignettes/namespace.html
* https://r-pkgs.org/dependencies-mindset-background.html#sec-dependencies-imports-vs-depends
* ?`::`: "The package namespace will be loaded if it was not loaded
  before the call, but the package will not be attached to the search
  path."